requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
slowapi==0.1.9
python-json-logger==2.0.7
aiohttp>=3.9.0

chromadb==0.4.18
PyPDF2==3.0.1
python-docx==1.1.0

alibabacloud_ice20201109>=2.0.1
alibabacloud_tea_openapi>=0.3.8
